% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcode.R
\name{QrNet}
\alias{QrNet}
\title{QrNet Function}
\usage{
QrNet(
  A,
  X,
  y,
  l = NULL,
  sigma = 1,
  tau = 0.5,
  tol = 1e-04,
  max_iter = 200,
  weights = rep(1, length(y)),
  verbose = 0,
  ...
)
}
\arguments{
\item{A}{A matrix or data frame representing the design matrix.}

\item{X}{An optional matrix of covariates, where `ncol(X)` is the number of predictors.}

\item{y}{A vector of the response variable.}

\item{l}{Regularization parameter (lambda) for the penalty term. If `NULL`, it is estimated from the data.}

\item{sigma}{A parameter controlling the strength of the regularization (default is 1).}

\item{tau}{Quantile level for regression (default is 0.5 for median regression).}

\item{tol}{Tolerance for convergence in the ADMM optimization algorithm (default is 1e-4).}

\item{max_iter}{Maximum number of iterations for the ADMM algorithm (default is 200).}

\item{weights}{A vector of weights for the observations (default is equal weights).}

\item{verbose}{If non-zero, provides verbose output during function execution.}

\item{...}{Additional arguments passed to the lower-level functions.}
}
\value{
A list of results from the quantile regression, including estimated coefficients, residuals, and other diagnostics.
}
\description{
This function implements a method for Quantile Regression using ADMM (Alternating Direction Method of Multipliers).
}
\examples{
\dontrun{
# Example of usage with simulated data:
A <- matrix(rnorm(100), nrow = 20)
X <- matrix(rnorm(40), nrow = 20)
y <- rnorm(20)
result <- QrNet(A, X, y)
print(result) 
}
}
