\name{SA24204148-package}
\alias{SA24204148-package}
\alias{SA24204148}
\docType{package}
\title{
  Network-Assisted Quantile Regression
}
\description{
  The package is useful for applications where both network structure and quantile regression are important, such as finance, social network analysis, and econometrics.
}
\details{
   The package implements a novel method for network-assisted quantile regression. It combines network structure data with quantile regression to estimate conditional quantiles while accounting for the latent network variables. The package provides two main algorithms: two_stage, a two-step method that first extracts the latent network variables using singular value decomposition (SVD), and _QrNet_, an augmented Lagrangian-based algorithm that simultaneously estimates the regression coefficients and latent network variables.
}
\author{
XintingHuang

Maintainer: XintingHuang <1792528405@qq.com>
}
\keyword{ package }
